DON'T MERGE! This is just to make a build for diagnostic purposes.
Not to be confused with the same-named branch in other repo's (CuraEngine and libArcus) which very much _should_ be merged. Look, if you're confused by this message -- it made sense at the time, and if this is in the main branch or any version-branches then I probably don't need to convince you to hecking revert it already right now.

done as part of CURA-12675