Disable USB-cable-printing by default (except for people who used it already).
We set the preference value for people who used USB-printing in the previous version, from the next version onwards, it'll be disabled by default -- as not many people still print by USB-cable, and more importantly, it can mess up some systems. It's not going away at the moment though.

CURA-12333