Update AddCoolingProfile.py
Added "Enable" setting so it can be left installed, but not run.
Added "Jump Start Fan" as an option when "By Feature" is selected.  Gets the fan spinning just before it is called.

Update AddCoolingProfile.py

Added "Enable" setting so it can be left installed, but not run.
Added "Jump Start Fan" as an option when "By Feature" is selected.  Gets the fan spinning just before it is called.