Update AnnealingOrDrying.py
The "drying" gcode turned out to be unsuitable for UM machines.  This commit is an attempt to rectify that by altering some lines in the "startup" and adding "SKIP_PROCEDURES".
There may still be an "incompatibility" warning on the printer but it can be ignored and the drying gcode will run.

Update AnnealingOrDrying.py

Oops.